Creación del caso de uso 1
Se implementó el Registro de Usuarios

Entidades involucradas: USUARIO, ROL
1.	(Controlador) Endpoints POST /api/usuarios, GET /api/usuarios, PUT /api/usuarios/{id}, DELETE /api/usuarios/{id} en UsuarioController.
2.	(DTOs) Crear app/cafeteria/dtos/user con: nombre, usuario, email, rol, activo, etc.).
3.	(Services) Métodos en UsuarioService/UsuarioServicesImpl: crearUsuario, listarUsuarios, editarUsuario, eliminarUsuario.
4.	(Reglas de Negocio) Validar que usuario/email/dpi sean únicos.
5.	(Seguridad) Guardar contrasenia_hash encriptado con argon2id.
6.	(Persistencia) UsuarioRepository y relación con RolRepository

+++ src/main/java/com/app/cafeteria/Repositories/CategoriaRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.app.cafeteria.entities.Categoria;
 
 @Repository
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> {}
+public interface CategoriaRepository extends JpaRepository<Categoria, Integer> {}
