+++ src/main/java/com/app/cafeteria/Repositories/PedidoRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.app.cafeteria.entities.Pedido;
 
 @Repository
-public interface PedidoRepository extends JpaRepository<Pedido, Long> {}
+public interface PedidoRepository extends JpaRepository<Pedido, Integer> {}
