+++ src/main/java/com/app/cafeteria/Repositories/ProductoRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.app.cafeteria.entities.Producto;
 
 @Repository
-public interface ProductoRepository extends JpaRepository<Producto, Long> {}
+public interface ProductoRepository extends JpaRepository<Producto, Integer> {}
